extracted title and background from index slides
