+++ pages/index.js
@@ -2,6 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import Page from '../components/Page';
+import SlideBackground from '../components/index/SlideBackground';
+import SlideTitle from '../components/index/SlideTitle';
 
 const maxProgress = 100;
 
@@ -19,121 +21,11 @@ const SlideContainer = styled.div`
     background-color: #fff;
 `;
 
-const SlideTitle = styled.header`
-    color: #231F20;
-    font-family: Teko;
-    text-align: center;
-    text-transform: uppercase;
-
-    position: fixed;
-    top: 50%;
-    left: 50%;
-
-    transform: translate(-50%, -50%);
-
-    > h1 {
-        font-size: 64px;
-        font-weight: 100;
-        line-height: 64px;
-    }
-
-    > h2 {
-        font-size: 28px;
-        font-weight: 100;
-        line-height: 40px;
-    }
-`;
-
-const SlideBackground = styled.div`
-    position: fixed;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-
-    background-image: url(${(props) => props.image});
-    background-size: cover;
-
-    will-change: transform, filter;
-`;
-
 const MountainCloud = styled.img`
     position: fixed;
 
     will-change: transform, opacity;
 `;
-
-class SlideTitleContainer extends React.Component {
-    static propTypes = {
-        animationProgress: PropTypes.number.isRequired,
-        progressStep: PropTypes.number.isRequired,
-        fadeIn: PropTypes.bool,
-        children: PropTypes.node,
-    };
-
-    static defaultProps = {
-        fadeIn: true,
-    };
-
-    render() {
-        return (
-            <SlideTitle style={{
-                opacity: this.getOpacity(this.props.animationProgress, this.props.progressStep, this.props.fadeIn),
-            }}>
-                {this.props.children}
-            </SlideTitle>
-        );
-    }
-
-    getOpacity = (animationProgress, progressStep, fadeIn = true) => {
-        const center = progressStep / 2;
-
-        if (animationProgress < center && !fadeIn) {
-            return 1;
-        }
-
-        const diff = Math.abs(animationProgress % progressStep - center);
-
-        return 1 - diff / center;
-    };
-}
-
-class SlideBackgroundContainer extends React.Component {
-    static propTypes = {
-        animationProgress: PropTypes.number.isRequired,
-        image: PropTypes.string.isRequired,
-        animationSplit: PropTypes.number.isRequired,
-        topOpacity: PropTypes.number.isRequired,
-        startOpacity: PropTypes.number.isRequired,
-    };
-
-    render() {
-        return (
-            <SlideBackground
-                image={this.props.image}
-                style={{
-                    transform: 'scale(' + (1 + this.props.animationProgress / 500) + ')',
-                    opacity: this.getBackgroundOpacity(),
-                }} />
-        );
-    }
-
-    getBackgroundOpacity = () => {
-        const increaseOpacity = this.props.topOpacity - this.props.startOpacity;
-
-        if (this.props.animationProgress <= this.props.animationSplit) {
-            return this.props.startOpacity + (
-                increaseOpacity * this.props.animationProgress / this.props.animationSplit
-            );
-        }
-
-        return this.props.topOpacity - (
-            this.props.topOpacity * (
-                (this.props.animationProgress - this.props.animationSplit) / (maxProgress - this.props.animationSplit)
-            )
-        );
-    };
-}
 
 class MountainCloudContainer extends React.Component {
     static propTypes = {
@@ -172,36 +64,36 @@ class MountainSlide extends React.PureComponent {
             // empty slide for transition of background images
         } else if (this.props.animationProgress > 2 * this.progressStep) {
             slide = (
-                <SlideTitleContainer animationProgress={this.props.animationProgress} progressStep={this.progressStep}>
+                <SlideTitle animationProgress={this.props.animationProgress} progressStep={this.progressStep}>
                     <h1>Learn, talk and ski</h1>
-                </SlideTitleContainer>
+                </SlideTitle>
             );
         } else if (this.props.animationProgress > this.progressStep) {
             slide = (
-                <SlideTitleContainer animationProgress={this.props.animationProgress} progressStep={this.progressStep}>
+                <SlideTitle animationProgress={this.props.animationProgress} progressStep={this.progressStep}>
                     <h2>25 - 28 January 2018</h2>
                     <h1>
                         Experts and industry leaders will
                         come together to showcase their work
                         in ReactJS, React Native and more
                     </h1>
-                </SlideTitleContainer>
+                </SlideTitle>
             );
         } else {
             slide = (
-                <SlideTitleContainer
+                <SlideTitle
                     animationProgress={this.props.animationProgress}
                     progressStep={this.progressStep}
                     fadeIn={false}>
                     <h2>The international event for coding inspiration</h2>
                     <h1>AgentConf 2018</h1>
-                </SlideTitleContainer>
+                </SlideTitle>
             );
         }
 
         return (
             <SlideContainer>
-                <SlideBackgroundContainer
+                <SlideBackground
                     animationProgress={this.props.animationProgress}
                     animationSplit={90}
                     startOpacity={0.4}
@@ -241,9 +133,9 @@ class CitySlide extends React.PureComponent {
             // empty slide
         } else if (this.props.animationProgress > this.progressStep) {
             slideTitleContainer = (
-                <SlideTitleContainer animationProgress={this.props.animationProgress} progressStep={this.progressStep}>
+                <SlideTitle animationProgress={this.props.animationProgress} progressStep={this.progressStep}>
                     <h1>Dornbirn, Austria</h1>
-                </SlideTitleContainer>
+                </SlideTitle>
             );
         } else {
             // empty slide
@@ -251,7 +143,7 @@ class CitySlide extends React.PureComponent {
 
         return (
             <SlideContainer>
-                <SlideBackgroundContainer
+                <SlideBackground
                     animationProgress={this.props.animationProgress}
                     animationSplit={50}
                     startOpacity={0.0}
@@ -283,6 +175,7 @@ export default class Index extends React.PureComponent {
         window.removeEventListener('resize', this.updateWindowHeight);
         window.removeEventListener('scroll', this.updateScrollY);
     }
+
     render() {
         const mountainSlideHeight = maxProgress * this.mountainSlideScrollDividend;
         const citySlideHeight = maxProgress * this.citySlideScrollDividend + this.state.windowHeight;
